+++ pytsite/__main__.py
@@ -1,9 +1,10 @@
-"""PytSite __main__ Module.
+"""PytSite __main__ Module
 """
+from pytsite import console
+
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
-from pytsite import console as _console
 
-_console.run()
+console.run()