+++ pytsite/contact/__init__.py
@@ -12,6 +12,7 @@ def __init():
     from pytsite import assetman, lang, tpl, browser
 
     assetman.register_package(__name__)
+    assetman.add('pytsite.contact@js/common.js', forever=True)
     lang.register_package(__name__)
     tpl.register_package(__name__)
     browser.register_ep('pytsite.contact.ep.submit')